LayoutTests/platform/ios-simulator/editing/pasteboard/paste-code-in-pre-expected.txt

This tests a bug when copying HTML markup inside <pre> tags. When pasted, this content would appear as the rendered form of that markup.
You should see markup in the editable region below. See <rdar://5027857>.
| <pre>
|   contenteditable="true"
|   id="pre"
|   <font>
|     face="Times"
|     <span>
|       style="white-space: normal; -webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"
|       "<input type='button'>foo<br>bar<b>baz</b><#selection-caret>"
